Dataframe of inputs and outputs of a health economic model developed and evaluated with the iviRA R package for testing

A dataframe containing 1,000 sets of probabilistic inputs and outputs of a individual-level health state transition model developed using iviRA package available on Github.

Usage

df_iviRA_pa

Format

See the documentation of the `sample_pars` function of the iviRA package for details.

Source

Incerti D, Curtis JR, Shafrin J, Lakdawalla DN, Jansen JP (2019). “ A Flexible Open-Source Decision Model for Value Assessment of Biologic Treatment for Rheumatoid Arthritis.” PharmacoEconomics. doi: 10.1007/s40273-018-00765-2.

Link to Github repository: https://innovationvalueinitiative.github.io/IVI-RA/index.html

Link to iviRA website: https://innovationvalueinitiative.github.io/IVI-RA/